Troca os f64 por Decimal.

+++ src/politica.rs
@@ -1,4 +1,5 @@
 use crate::filtros::Filtro;
+use rust_decimal::Decimal;
 
 #[derive(Clone, Debug, Default)]
 pub struct Politica {
@@ -11,13 +12,13 @@ pub struct Regra {
     pub editavel: bool,
     pub aplicavel_a_promocoes: bool,
     pub filtro: Box<dyn Filtro>,
-    pub desconto_maximo: f64,
-    pub desconto_sugerido: f64,
+    pub desconto_maximo: Decimal,
+    pub desconto_sugerido: Decimal,
 }
 
 #[derive(Copy, Clone, Debug, Default)]
 pub struct RegraItemPedido {
     pub regra_id: u32,
     pub item_id: u32,
-    pub desconto: f64,
+    pub desconto: Decimal,
 }
